[ADDED] Open weather logo in the into screen
